# Conflicts:
#	OnlineBookstoreManagementSystem.iml

+++ src/main/java/com/example/pojo/User.java
@@ -4,13 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
-
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class User {
-    private BigInteger userId;
+    private long userId;
     private String userName;
     private String userPassword;
     private String address;
